One big push

+++ backend/routers/users.py
@@ -4,9 +4,10 @@ from sqlalchemy import or_, and_
 
 from backend.database import get_db
 from backend.db_models.users import Users
-from backend.pydantic_models.users import UserBase, User
+from backend.pydantic_models.users import UserBase, User, UserLogin
 from backend.db_models.sessions import UserSession
 from backend.services.dependencies import set_login_session_id, is_authenticated
+from backend.services.helpers import delete_urls
 
 
 router = APIRouter(
@@ -16,25 +17,34 @@ router = APIRouter(
 
 
 @router.post("/login")
-async def login(user: UserBase, request: Request, response: Response, db: Session = Depends(get_db)):
-    user = db.query(Users).filter(and_(Users.username == user.username, Users.password == user.password)).first()
+async def login(user: UserLogin, request: Request, response: Response, db: Session = Depends(get_db)):
+    user = db.query(Users).filter(
+        and_(
+            or_(
+                Users.username == user.username_or_email,
+                Users.email == user.username_or_email
+            ),
+            Users.password == user.password
+        )
+    ).first()
+
     if not user:
         raise HTTPException(status_code=401, detail="Invalid credentials")
-    print("User found", user)
     user_session = set_login_session_id(request, response, user, db=db)
-    print("User session", user_session)
     return User.model_validate(user)
 
 @router.post("/")
-async def create_user(user: UserBase, db: Session = Depends(get_db)):
+async def create_user(user: UserBase, request: Request, response: Response, db: Session = Depends(get_db)):
     existing_user = db.query(Users).filter(or_(Users.username == user.username, Users.email == user.email)).first()
     if existing_user:
         raise HTTPException(status_code=400, detail="User already exists")
     new_user = Users(username=user.username, password=user.password, email=user.email)
     db.add(new_user)
     db.commit()
-    return {"message": "User created successfully"}
+    user_session = set_login_session_id(request, response, new_user, db=db)
+    return User.model_validate(new_user)
 
+# this also acts as a check for the authentication
 @router.get("/", response_model=User)
 async def get_user(user_session: UserSession = Depends(is_authenticated)):
     user = user_session.user
@@ -47,6 +57,9 @@ async def delete_user(request: Request, response: Response, db: Session = Depend
     user = user_session.user
     if not user:
         raise HTTPException(status_code=404, detail="User not found")
+    success = delete_urls(user, db=db)
+    if not success:
+        raise HTTPException(status_code=500, detail="Failed to delete user")
     db.delete(user)
     db.delete(user_session)
     db.commit()
